chore: extract shared logic from trin-execution::evm into trin-evm
