+++ .circleci/config.yml
@@ -232,7 +232,7 @@ jobs:
     # parallelism level should be set to the amount of simulators we have or greater
     # The reason for this is the CI code currently only supports 1 input at a time
     # if we have a parallelism level of 5 and 6 sims one test runner will get 2 test sims and fail
-    parallelism: 17
+    parallelism: 18
     steps:
       - checkout
       - run:
