+++ portal-bridge/Cargo.toml
@@ -46,6 +46,7 @@ serde_yaml.workspace = true
 ssz_types.workspace = true
 tokio.workspace = true
 tracing.workspace = true
+trin-evm.workspace = true
 trin-execution.workspace = true
 trin-metrics.workspace = true
 trin-utils.workspace = true
