+++ trin-evm/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "trin-evm"
+version = "0.1.0"
+edition = "2021"
+repository = "https://github.com/ethereum/trin/tree/master/trin-evm"
+license = "GPL-3.0"
+readme = "README.md"
+keywords = ["ethereum", "execution-layer"]
+categories = ["cryptography::cryptocurrencies"]
+description = "Common EVM functionality"
+authors = ["https://github.com/ethereum/trin/graphs/contributors"]
+
+
+[dependencies]
+alloy-primitives.workspace = true
+alloy-rlp.workspace = true
+alloy-rpc-types.workspace = true
+ethportal-api.workspace = true
+revm.workspace = true
+revm-primitives.workspace = true
+tokio.workspace = true
