+++ trin-evm/README.md
@@ -1,0 +1,3 @@
+# Trin EVM
+
+The collection of common EVM related functionality.
