+++ trin-execution/Cargo.toml
@@ -40,6 +40,7 @@ tempfile.workspace = true
 thiserror.workspace = true
 tokio.workspace = true
 tracing.workspace = true
+trin-evm.workspace = true
 trin-utils.workspace = true
 
 [dev-dependencies]
