+++ trin-execution/src/era/binary_search.rs
@@ -5,8 +5,7 @@ use e2store::{
 };
 use reqwest::Client;
 use revm_primitives::SpecId;
-
-use crate::evm::spec_id::get_spec_block_number;
+use trin_evm::spec_id::get_spec_block_number;
 
 use super::{
     constants::FIRST_ERA_EPOCH_WITH_EXECUTION_PAYLOAD,
