+++ trin-execution/src/evm/block_executor.rs
@@ -20,6 +20,9 @@ use revm::{
 use revm_primitives::{keccak256, Address, ResultAndState, SpecId, B256, U256};
 use serde::{Deserialize, Serialize};
 use tracing::info;
+use trin_evm::{
+    create_block_env, create_evm_with_tracer, spec_id::get_spec_id, tx_env_modifier::TxEnvModifier,
+};
 
 use crate::{
     era::types::{ProcessedBlock, TransactionsWithSender},
@@ -30,11 +33,7 @@ use crate::{
     storage::evm_db::EvmDB,
 };
 
-use super::{
-    create_block_env, create_evm_with_tracer,
-    post_block_beneficiaries::get_post_block_beneficiaries, spec_id::get_spec_id,
-    tx_env_modifier::TxEnvModifier,
-};
+use super::post_block_beneficiaries::get_post_block_beneficiaries;
 
 pub const BLOCKHASH_SERVE_WINDOW: u64 = 256;
 const GENESIS_STATE_FILE: &str = "trin-execution/resources/genesis/mainnet.json";
